cmd/contour: clean up leadership election setup
Only register leadership election workers if leadership election is
enabled.

Also remove the event recorder and broadcaster because it doesn't seem
to do anything -- leadership election still poops klog lines.
